requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.2
jupyterlab-git==0.20.0
jupyterlab-server==1.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
moviepy==1.0.3
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
proglog==0.1.9
progressbar2==3.51.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tifffile==2020.5.11
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
-e git+https://github.com/dlpbc/keras-kinetics-i3d.git#egg=inception3d